feat: don’t expose master mnemonic
As a first phase of protecting the master keypair, this change moves it to the rust side and only exposes derived keys and accounts. Phase 2 should be to move signing to backend as well, and keep keys on backend apart from the bot key.

Also fixes the overlays that were not showing because they had a top level await.