feat: adaptable mining frames
The mining frames now last a set number of reward ticks, which will make the frames themselves a bit more dynamic in how long they last.

Mining frames are now synched prior to a build and then updated once an app is launched. The frames are updated in the background on start.