feat: improve installer speed
This change hashes docker configs so it can skip force rebuild unless absolutely needed.

As part of this, also kill a hung installer if running on a previous version, and use the installer lock as the pid source, not pgrep.

Finally, the bitcoin docker now mounts the bitcoin configs instead of embedding them so that we don’t need to detect changes, and the miner removes it’s unneeded docker.